+++ NEWS.rst
@@ -7,6 +7,9 @@ Changes in 5.15 (released ??/??/2015)
 *	The PySQL command ``setvar`` now uses the ``name`` key as the variable name
 	instead of the ``var`` key.
 
+*	A new PySQL command ``unsetvar`` has been added for deleting an existing
+	variable.
+
 
 Changes in 5.14.1 (released 12/04/2015)
 ---------------------------------------